refactor: Sprint 48 PR 3 — split telegram.rs adapter (3b)
Extract 6 sub-modules from telegram/mod.rs (adapter + outbound):
- creds.rs (64 LOC): credential resolution
- reply.rs (433 LOC): reply/provenance helpers + 7 tests
- bot_api.rs (150 LOC): react/edit/download + 3 tests
- notify.rs (99 LOC): notification helpers
- adapter.rs (747 LOC): TelegramChannel + Channel/UxEventSink impls + 15 tests
- bootstrap.rs (205 LOC): init_from_config, resolve_fleet_binding

mod.rs reduced to 31 LOC (re-exports only).

adapter.rs is 747 LOC (47 over 700 target) — production code is 440 LOC,
excess is from cargo fmt expanding compact test constructors. All other
sub-files well under 700.

Zero behavior change. 1419 tests = baseline.
